requirements/development.txt
-r common.txt
ipdb
pudb
werkzeug
django-werkzeug-debugger-runserver
django-dynamic-fixture
